refactor: 데이터 패칭 로직을 App 컴포넌트로 이동
