+++ src/stores/editorStore.js
@@ -32,6 +32,11 @@ export default class EditorStore {
     return { document: recentDocument, documentId, isLocalData: recentDocument === localDocument };
   }
 
+  /**
+   * 스토어의 상태에 저장된 문서 정보를 로컬 스토리지에 저장하고, 서버에 디바운스로 업데이트 요청을 합니다.
+   * @param {function} onSuccess 문서 저장 성공 시 실행할 콜백 함수
+   * @param {number} timeout 디바운스 시간
+   */
   saveDocument(onSuccess = () => {}, timeout = 1000) {
     const { documentId, document } = this.state;
     const { title, content } = document;
@@ -46,10 +51,14 @@ export default class EditorStore {
       await modifyDocument(document, documentId);
       storage.removeItem(DOCUMENT(documentId));
       onSuccess();
-    }, 10000);
+    }, timeout);
   }
 
-  // TODO: 스토리지에 존재하는 모든 데이터 읽어서 최신일 경우 서버에 modify 요청하기.
+  /**
+   * 로컬 스토리지에 존재하는 모든 데이터를 읽고, 서버에서 가져온 문서보다 최신 정보인 경우에 서버에 업데이트 요청을 합니다.
+   * @param {array} documents 서버에서 가져온 문서 목록
+   * @returns
+   */
   pushStorageDocuments(documents = []) {
     const promises = Object.keys(localStorage)
       .filter((key) => key.startsWith(DOCUMENT()))
